requirements.txt
asgiref==3.2.3
attrs==19.3.0
Django==3.0.4
importlib-metadata==1.5.0
more-itertools==8.2.0
numpy==1.18.1
packaging==20.3
pkg-resources==0.0.0
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
python-dateutil==2.8.1
pytz==2019.3
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scipy==1.4.1
-e git+https://github.com/NTIA/SigMF.git@multi-recording-archive#egg=SigMF
six==1.14.0
sqlparse==0.3.1
wcwidth==0.1.8
zipp==3.1.0
